refactor:common ui 폴더 정리
