+++ app/(pages)/(auth)/login/email/page.tsx
@@ -6,7 +6,7 @@ import Button from "@/common/components/ui/buttons/Button/Button";
 import Link from "next/link";
 import UseForm from "@/feature/auth/hooks/Form/useForm";
 import { LoginFormState } from "@/feature/auth/type";
-import ArrowBackTopBar from "@/common/components/ui/topBar/ArrowBackTopBar/ArrowBackTopBar";
+import ArrowBackTopBar from "@/common/components/ui/topBar/ArrowBackTopBar";
 import TextWithBorder from "@/feature/auth/components/Login/molecules/TextWithBorder";
 
 export default function LoginEmailPage() {
