+++ app/(pages)/(auth)/register/page.tsx
@@ -5,7 +5,7 @@ import Button from "@/common/components/ui/buttons/Button/Button";
 import UseForm from "@/feature/auth/hooks/Form/useForm";
 import { LoginFormState, RegisterFormState } from "@/feature/auth/type";
 import { useRouter } from "next/navigation";
-import ArrowBackTopBar from "@/common/components/ui/topBar/ArrowBackTopBar/ArrowBackTopBar";
+import ArrowBackTopBar from "@/common/components/ui/topBar/ArrowBackTopBar";
 import useFetching from "@/common/hooks/useFetching";
 import LoadingUI from "@/common/components/ui/loading/LoadingUI";
 import PageDarkWrapper from "@/common/components/ui/wrapper/PageDarkWrapper";
