+++ app/(pages)/(withFooter)/record/search/page.tsx
@@ -1,4 +1,4 @@
-import ArrowBackTopBar from "@/common/components/ui/topBar/ArrowBackTopBar/ArrowBackTopBar";
+import ArrowBackTopBar from "@/common/components/ui/topBar/ArrowBackTopBar";
 import PlaceInfoCard from "@/feature/place/components/PlaceInfo/organisms/PlaceInfoCard";
 import TextSearchBar from "@/feature/search/components/SearchText/molecules/TextSearchBar";
 import { getTextSearchPlaceData } from "@/feature/search/queries/getTextSearchPlaceData";
