+++ app/common/components/ui/topBar/ArrowBackTopBar.tsx
@@ -2,7 +2,7 @@
 
 import { useRouter } from "next/navigation";
 import { twMerge } from "tailwind-merge";
-import { TopBarProps } from "../type";
+import { TopBarProps } from "./type";
 import ArrowBackIcon from "@/common/assets/icons/arrow/ArrowBackIcon";
 
 //Molecule
