+++ app/common/components/ui/topBar/PageTopBar.tsx
@@ -2,7 +2,7 @@
 
 import MenuIcon from "@/common/assets/icons/menu/MenuIcon";
 import { twMerge } from "tailwind-merge";
-import Title from "../../text/Title";
+import Title from "../text/Title";
 import { useSetRecoilState } from "recoil";
 import { isModalOpenSelector } from "@/common/state/modal-open";
