+++ app/feature/place/components/PlaceDetail/organisms/PlaceDetailTopBar.tsx
@@ -5,7 +5,7 @@ import { copyLink } from "@/common/utils/text/copy-link";
 import { usePathname } from "next/navigation";
 import { useSetRecoilState } from "recoil";
 import { toastInfoSelector } from "@/common/state/toast";
-import ArrowBackTopBar from "@/common/components/ui/topBar/ArrowBackTopBar/ArrowBackTopBar";
+import ArrowBackTopBar from "@/common/components/ui/topBar/ArrowBackTopBar";
 import UseMap from "@/common/components/ui/map/Map/useMap";
 import { PlaceDetailInfoProps } from "@/feature/place/queries/dto/place-detail";
 import MapIcon from "@/common/assets/icons/map/mapIcon";
